Integrate Razorpay test payments with checkout

+++ Backend/order-service/src/main/java/com/cropdeal/orderservice/dto/PaymentProcessingResponse.java
@@ -8,8 +8,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class PaymentProcessingResponse {
-	private String apikey;
-	private String currency;
+    private String apikey;
+    private String currency;
     private String orderId;
-    private double amount;
+    private long amount;
 }
